+++ src/StackExchange.Redis/RedisLiterals.cs
@@ -110,6 +110,7 @@ namespace StackExchange.Redis
             TYPE = "TYPE",
             WEIGHTS = "WEIGHTS",
             WITHSCORES = "WITHSCORES",
+            WITHVALUES = "WITHVALUES",
             XOR = "XOR",
             XX = "XX",